base de autenticacao funcionando
falta adicionar a regra de negocio: se o email e a senha forem validos entao o login funciona
